feat: build error fix and refactor main list page & overlays
