+++ packages/service-frontend/app/components/index.ts
@@ -1,2 +1,3 @@
 export * from './common'
+export * from './marketplace/deck'
 export * from './marketplace/game'
