+++ packages/service-frontend/app/components/marketplace/deck/index.ts
@@ -1,0 +1,3 @@
+export { default as AllDeckInitialItem } from './AllDeckInitialItem'
+export { default as AllDeckLastItem } from './AllDeckLastItem'
+export { default as MyDeckList } from './MyDeckList'
