+++ packages/service-frontend/app/components/overlay/constants.ts
@@ -1,0 +1,4 @@
+export enum OverlayType {
+  CreateDeck = 'CreateDeck',
+  Onboarding = 'Onboarding',
+}
