+++ packages/service-frontend/app/components/overlay/index.ts
@@ -1,0 +1,2 @@
+export { default as CreateDeckOverlay } from './CreateDeckOverlay'
+export { default as OnboardingOverlay } from './OnboardingOverlay'
